front(D): review round 2 — log type track, token-resolution doctrine correction, shared chart utility
