+++ docs/ARCHITECTURE.md
@@ -630,7 +630,7 @@ A mounted declarative widget owns its timers, abort controllers, chart objects, 
 Dashboard groups Logs, Evolution, Costs, Updates, and Activity under one page. Sub-tabs activate their own loading and refresh policy instead of running every expensive reader continuously while hidden. Logs and Chat share event summarization (`log_events.js`) so task phases are described consistently. The same `taskOutcomeSeverity` reducer drives both surfaces: a degraded review or best-effort/degraded objective is warning-colored, never presented as a green solved result. Compact review projections disclose every panel and actor's model/provider, role, transport/parse/semantic state, coverage, quorum contribution, reason, and enforcement impact by default; since v6.70.0 the actor/panel `reason` is the COMPLETE redacted rationale (the former 500/800-char caps destroyed the only owner-reachable copy — reviewer rationale is a cognitive artifact, BIBLE P1) and each actor carries a forensic `response_ref` into the private observability store; full model prompts/responses themselves remain in that private audit storage. Evolution reads `/api/evolution-data`; Costs reads the physical-attempt ledger projection from `/api/cost-breakdown` and distinguishes confirmed/settled, reserved, unresolved upper bound, and unknown/unmetered work together with `cost_final`; it exposes budget controls from the shared setup contract via `/api/settings` GET/POST. Updates exposes official managed updates plus local recovery commits/tags.
 The Logs page renders task/LLM/tool/progress events as grouped task cards, while Chat renders the same stream as a live task card so operational history and live dialogue stay visually consistent.
 Worker tasks forward their `append_jsonl` log lines to the live dashboard over `EVENT_Q` via a per-worker log sink installed in `supervisor/workers.py::worker_main` (the WS log sink only exists in the main process), suppressing types that already arrive through a dedicated live sibling event — `tool_call`/`llm_round`/`task_checkpoint`/`task_done`/`llm_usage` — to avoid double broadcast and a double `task_checkpoint` file write. `EVENT_Q` belongs to the `server.py` process, not to a replaceable worker-pool generation: direct chat, consciousness, active turns, and every respawned worker retain one manager-backed queue proxy until process exit. The manager transport serializes in the producer and isolates producer connections, avoiding both feeder-thread false positives and pipe corruption when a worker is force-killed mid-frame. `spawn_workers` therefore reuses that bus, holds the lifecycle lock through complete pool publication, and refuses to overwrite a non-empty live pool; single-slot recovery uses `respawn_worker`. On load and on every reconnect the Logs page backfills recent history from `/api/logs/{events,tools,progress,supervisor}` (`web/modules/logs.js::backfillRecentLogs`) and dedupes the live-overlap window by event identity so the pre-connect window is neither dropped nor shown twice.
-Chart.js is bundled locally as `web/chart.umd.min.js`; no CDN dependency by design. Because a canvas cannot resolve CSS `var(...)`, `web/modules/utils.js::readThemeTokens` is the single canvas↔token boundary: the chart-theme helpers `evolution.js::evolutionChartTheme` and `widgets.js::chartPalette` (with `evolution.js::chartColorAlpha` for the dataset fills) resolve every series colour, axis colour and font stack off the live `:root` computed style at chart-build time, so no chart holds a hand-synced copy of the design tokens — and the token names they pass must be LEAF values, since an alias would come back as the literal string `var(...)` (pinned by `tests/test_widgets_ui_static.py::test_canvas_palettes_read_only_leaf_root_tokens`).
+Chart.js is bundled locally as `web/chart.umd.min.js`; no CDN dependency by design. Because a canvas cannot resolve CSS `var(...)`, `web/modules/utils.js::readThemeTokens` is the single canvas↔token boundary: the chart-theme helpers `evolution.js::evolutionChartTheme` and `widgets.js::chartPalette` (with `utils.js::chartColorAlpha` for the dataset fills — it sits beside `readThemeTokens` because both chart modules share it, rather than Widgets importing a colour utility from the Evolution page) resolve every series colour, axis colour and font stack off the live `:root` computed style at chart-build time, so no chart holds a hand-synced copy of the design tokens — and every token name they pass must be DEFINED in `:root`, because `getPropertyValue` returns the COMPUTED value: an alias such as `--accent-system: var(--amber)` resolves through to the hex and is safe to name, while an UNDEFINED custom property returns the empty string, which a canvas accepts and then silently drops the series for (pinned by `tests/test_widgets_ui_static.py::test_canvas_palettes_read_only_defined_root_tokens`).
 
 Logs merges live WebSocket log frames with bounded REST backfill from events, tools, progress, and supervisor logs. It orders the merged backfill chronologically, deduplicates overlap with the live stream and reconnect backfills, groups related task events into bounded cards, exposes the raw record on demand, and applies shared category/severity/review presentation from `log_events.js`. Clearing the visible panel does not delete the underlying logs.
